patch: some refactors and code cleanup
