+++ src/StockportWebapp/Models/CookieConsentLevel.cs
@@ -1,0 +1,17 @@
+﻿namespace StockportWebapp.Models;
+
+public class CookieConsentLevel
+{
+    [JsonProperty(PropertyName = "strictly-necessary")]
+    public bool StriclyNecessary { get; set; } = true;
+
+    [JsonProperty(PropertyName = "functionality")]
+    public bool Functionality { get; set; }
+
+    [JsonProperty(PropertyName = "tracking")]
+    public bool Tracking { get; set; }
+
+    [JsonProperty(PropertyName = "targeting")]
+    public bool Targetting { get; set; }
+
+}
