+++ src/StockportWebapp/Utils/CookiesHelper.cs
@@ -1,41 +1,10 @@
-﻿using StockportGovUK.NetStandard.Gateways.Models.Booking;
-
-namespace StockportWebapp.Utils;
-
-public interface ICookiesHelper
-{
-    void AddToCookies<T>(string slug, string cookieType);
-    List<string> GetCookies<T>(string cookieType);
-    List<T> PopulateCookies<T>(List<T> items, string cookieType);
-    void RemoveAllFromCookies<T>(string cookieType);
-    void RemoveFromCookies<T>(string slug, string cookieType);
-    CookieConsentLevel GetCurrentCookieConsentLevel();
-}
-
-public class CookieConsentLevel
-{
-    [JsonProperty(PropertyName = "strictly-necessary")]
-    public bool StriclyNecessary { get; set; } = true;
-
-    [JsonProperty(PropertyName = "functionality")]
-    public bool Functionality { get; set; }
-
-    [JsonProperty(PropertyName = "tracking")]
-    public bool Tracking { get; set; }
-
-    [JsonProperty(PropertyName = "targeting")]
-    public bool Targetting { get; set; }
-
-}
+﻿namespace StockportWebapp.Utils;
 
 public class CookiesHelper : ICookiesHelper
 {
     private IHttpContextAccessor httpContextAccessor;
 
-    public CookiesHelper(IHttpContextAccessor accessor)
-    {
-        httpContextAccessor = accessor;
-    }
+    public CookiesHelper(IHttpContextAccessor accessor) => httpContextAccessor = accessor;
 
     public List<T> PopulateCookies<T>(List<T> items, string cookieType)
     {
@@ -104,12 +73,8 @@ public class CookiesHelper : ICookiesHelper
         UpdateCookies(cookiesAsObject, cookieType);
     }
 
-    public List<string> GetCookies<T>(string cookieType)
-    {
-        var cookiesAsObject = GetCookiesAsObject(cookieType);
-
-        return cookiesAsObject.Values.SelectMany(cookieAsObject => cookieAsObject.Select(cookie => cookie.ToString())).ToList();
-    }
+    public List<string> GetCookies<T>(string cookieType) =>
+        GetCookiesAsObject(cookieType).Values.SelectMany(cookieAsObject => cookieAsObject.Select(cookie => cookie.ToString())).ToList();
 
     public static Dictionary<string, List<string>> ExtractValuesFromJson(string cookies)
     {
