+++ src/StockportWebapp/Utils/ICookiesHelper.cs
@@ -1,0 +1,11 @@
+﻿namespace StockportWebapp.Utils;
+
+public interface ICookiesHelper
+{
+    void AddToCookies<T>(string slug, string cookieType);
+    List<string> GetCookies<T>(string cookieType);
+    List<T> PopulateCookies<T>(List<T> items, string cookieType);
+    void RemoveAllFromCookies<T>(string cookieType);
+    void RemoveFromCookies<T>(string slug, string cookieType);
+    CookieConsentLevel GetCurrentCookieConsentLevel();
+}
